move entity list into own file and use index for exports
