+++ assets/src/components/ui/index.js
@@ -3,17 +3,15 @@ export { default as CalendarDateRange } from './dates-and-times/calendar-date-ra
 export { default as CalendarPageDate } from './dates-and-times/calendar-page-date';
 export { default as MediumCalendarDate } from './dates-and-times/medium-calendar-date';
 export { default as DropDownMenu } from './drop-down-menu';
-export {
-	EntityActionMenuItem,
-	useEntityActionMenuItems,
-} from './entity-actions-menu';
+export { EntityActionMenuItem, useEntityActionMenuItems }
+	from './entity-actions-menu';
 export { default as EntityDetailsPanel }
 	from './entity-details-panel/entity-details-panel';
-export { default as EntityList } from './entity-list';
 export { default as EntityPaperFrame } from './entity-paper-frame';
 export { default as EspressoButton } from './espresso-button';
 export { default as IconMenuItem } from './icon-menu-item/icon-menu-item';
 export { default as LoadingNotice } from './loading-notice';
+export * from './entity-list';
 export * from './entity-list-filter-bar';
 export * from './entity-pagination';
 export * from './image';
